requirements.txt
# biopython
# pandas
# numpy
# plotly
# matplotlib
# scikit-learn
# umap-learn


streamlit==1.28.0
pandas
numpy
plotly
matplotlib
biopython
scikit-learn
scipy
umap-learn
torch    # <-- if you need pytorch, consider specifying cpu-only wheel for mac/arm
watchdog
